server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.2.1
pytest==8.3.4
httpx==0.28.1
